Set up react navigation and initialize settings page.
