+++ package.json
@@ -11,6 +11,7 @@
     "react": "16.3.1",
     "react-native": "0.55.4",
     "react-native-firebase": "^4.2.0",
+    "react-navigation": "^2.0.4",
     "rsg-chess": "1.0.4",
     "rsg-chess-rn-graphics": "0.0.2"
   },
